ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
transformers
torch
transformers
sentence-transformers
faiss-cpu